refactor: extract doc lookup to object Doc
