+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/Help.scala
@@ -1,14 +1,11 @@
 package io.shiftleft.semanticcpg.language
 
 import io.shiftleft.codepropertygraph.generated.nodes.Node
-import io.shiftleft.semanticcpg.language.Help.StepDoc
 import io.shiftleft.semanticcpg.utils.Table
 import io.shiftleft.semanticcpg.{Doc, Traversal}
 import org.reflections.Reflections
 
 import scala.jdk.CollectionConverters._
-import scala.reflect.runtime.universe._
-import scala.tools.reflect.ToolBox
 
 object Help {
 
@@ -66,25 +63,9 @@ object Help {
     findStepDocs(classOf[NodeSteps[_]])
 
   private def findStepDocs(traversal: Class[_]): Iterable[StepDoc] =
-    docAnnotationByMethodname(traversal).map { case (methodName, doc) =>
+    Doc.docByMethodName(traversal).map { case (methodName, doc) =>
       StepDoc(traversal.getName, methodName, doc)
     }
-
-  private val mirror = runtimeMirror(this.getClass.getClassLoader)
-  private val mirrorToolbox = mirror.mkToolBox()
-
-  private def docAnnotationByMethodname(traversal: Class[_]): Iterable[(String, Doc)] = {
-    val traversalTpe = mirror.classSymbol(traversal).toType
-    def toDoc(annotation: Annotation): Doc =
-      mirrorToolbox.eval(mirrorToolbox.untypecheck(annotation.tree)).asInstanceOf[Doc]
-
-    traversalTpe.members
-      .filter(_.isPublic)
-      .map { member =>
-        val docAnnotationMaybe = member.annotations.filter(_.tree.tpe =:= typeOf[Doc]).map(toDoc).headOption
-        (member.name.toString, docAnnotationMaybe)
-      }.collect { case (methodName, Some(doc)) => (methodName, doc) }
-  }
 
   case class StepDoc(traversalClassName: String, methodName: String, doc: Doc)
 }
